requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.9.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
redis>=5.0.1
httpx>=0.25.2
numpy>=1.26.0
aiofiles>=23.2.1
python-dotenv>=1.0.0

# Optional: GPU/ML packages (uncomment if needed for actual GPU inference)
# torch>=2.1.0
# transformers>=4.35.0

